Preprocess Data

This function performs pre-processing steps to make expected points predictions on nflreadr data

Usage

ep_preprocess(pbp)

Arguments

pbp

pbp dataframe from nflreadr::load_pbp()

Value

a list of two dataframes (one for passes and one for rushes) of nflreadr data with the expectedpoints columns transformed for prediction

See also

Examples

# \donttest{
try({ # catch failures for CRAN purposes
  pbp_download <- readRDS(system.file("pbp_download.rds",package = "ffopportunity"))
  # this file is equivalent to nflreadr::load_pbp(2021) %>% head(1000)
  ep_preprocess(pbp_download)
  })
#> $rush_df
#> # A tibble: 252 × 43
#>    game_id  play_id desc     rusher_player_id full_name posteam two_point_attem…
#>    <chr>      <dbl> <chr>    <chr>            <chr>     <chr>              <dbl>
#>  1 2021_01…      55 (15:00)… 00-0032764       Derrick … TEN                    0
#>  2 2021_01…     181 (12:55)… 00-0035228       Kyler Mu… ARI                    0
#>  3 2021_01…     282 (11:40)… 00-0034681       Chase Ed… ARI                    0
#>  4 2021_01…     473 (9:17) … 00-0032764       Derrick … TEN                    0
#>  5 2021_01…     716 (8:11) … 00-0032764       Derrick … TEN                    0
#>  6 2021_01…     737 (7:24) … 00-0032764       Derrick … TEN                    0
#>  7 2021_01…     816 (6:44) … 00-0034681       Chase Ed… ARI                    0
#>  8 2021_01…    1011 (2:24) … 00-0032764       Derrick … TEN                    0
#>  9 2021_01…    1032 (1:49) … 00-0032764       Derrick … TEN                    0
#> 10 2021_01…    1185 (14:16)… 00-0033553       James Co… ARI                    0
#> # … with 242 more rows, and 36 more variables: two_point_converted <dbl>,
#> #   rush_attempt <dbl>, first_down_rush <dbl>, fumble_lost <dbl>, season <dbl>,
#> #   week <int>, rushing_yards <dbl>, rush_touchdown <fct>, first_down <fct>,
#> #   posteam_type <chr>, run_location <chr>, run_gap <chr>, run_gap_dir <chr>,
#> #   surface <chr>, wind <int>, temp <int>, roof <chr>, position <chr>,
#> #   yardline_100 <dbl>, half_seconds_remaining <dbl>,
#> #   game_seconds_remaining <dbl>, fixed_drive <dbl>, era <chr>, xpass <dbl>, …
#> 
#> $pass_df
#> # A tibble: 430 × 50
#>    game_id  play_id desc       passer_player_id passer_full_name passer_position
#>    <chr>      <dbl> <chr>      <chr>            <chr>            <chr>          
#>  1 2021_01…      76 (14:23) (… 00-0029701       Ryan Tannehill   QB             
#>  2 2021_01…     100 (13:42) (… 00-0029701       Ryan Tannehill   QB             
#>  3 2021_01…     152 (13:27) (… 00-0035228       Kyler Murray     QB             
#>  4 2021_01…     218 (12:26) (… 00-0035228       Kyler Murray     QB             
#>  5 2021_01…     253 (11:54) (… 00-0035228       Kyler Murray     QB             
#>  6 2021_01…     386 (10:08) (… 00-0035228       Kyler Murray     QB             
#>  7 2021_01…     410 (9:29) (N… 00-0035228       Kyler Murray     QB             
#>  8 2021_01…     599 (8:30) (S… 00-0035228       Kyler Murray     QB             
#>  9 2021_01…     621 (8:26) (S… 00-0035228       Kyler Murray     QB             
#> 10 2021_01…     660 (8:18) (S… 00-0035228       Kyler Murray     QB             
#> # … with 420 more rows, and 44 more variables: receiver_player_id <chr>,
#> #   receiver_full_name <chr>, receiver_position <chr>, posteam <chr>,
#> #   two_point_attempt <dbl>, two_point_converted <dbl>, pass_attempt <dbl>,
#> #   receiving_yards <dbl>, first_down_pass <dbl>, fumble_lost <dbl>,
#> #   season <dbl>, week <int>, complete_pass <fct>, yards_after_catch <dbl>,
#> #   pass_touchdown <fct>, first_down <fct>, interception <fct>,
#> #   relative_to_endzone <dbl>, wind <int>, score_differential <dbl>, …
#> 
# }